+++ src/prelude.rs
@@ -44,14 +44,6 @@
 
 pub use core::any::Any;
 
-#[cfg(feature = "anyhow")]
-pub use crate::compat::anyhow::IntoAnyhow;
-#[cfg(feature = "boxed-error")]
-pub use crate::compat::boxed_error::IntoBoxedError;
-#[cfg(feature = "error-stack")]
-pub use crate::compat::error_stack::IntoErrorStack;
-#[cfg(feature = "eyre")]
-pub use crate::compat::eyre::IntoEyre;
 pub use crate::{
     Report, bail, compat::IntoRootcause, handlers, iterator_ext::IteratorExt, markers, report,
     report_attachment, result_ext::ResultExt,